pyproject.toml
[tool.poetry]
name = "wdgram"
version = "0.1.0"
description = ""
authors = ["fauzaanu <fauzaangasim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.42.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
